requirements.txt
anyio==3.6.2
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==3.1.0
click==8.1.3
fastapi==0.95.0
googlesearch-python==1.1.0
h11==0.14.0
httptools==0.5.0
idna==2.10
pydantic==1.10.7
python-dotenv==1.0.0
PyYAML==6.0
requests==2.25.1
sniffio==1.3.0
soupsieve==2.4
starlette==0.26.1
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
